Move objects to the clinicalTrial namespace.
The ApiReference and Delighters business objects are specific to
Clinical Trial testing.

Addresses issues #30 and #32.
