+++ src/main/java/gov/nci/clinicalTrial/common/ApiReference.java
@@ -1,4 +1,4 @@
-package gov.nci.commonobjects;
+package gov.nci.clinicalTrial.common;
 
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
