+++ src/test/java/gov/nci/testcases/BasicSearch_Test.java
@@ -15,12 +15,11 @@ import org.testng.annotations.Test;
 
 import gov.nci.Utilities.BrowserManager;
 import gov.nci.Utilities.ExcelManager;
-//import gov.nci.clinicalTrial.pages.AdvanceSearch;
 import gov.nci.clinicalTrial.pages.BasicSearch;
-import gov.nci.commonobjects.ApiReference;
+import gov.nci.clinicalTrial.common.ApiReference;
+import gov.nci.clinicalTrial.common.Delighters;
 import gov.nci.commonobjects.Banner;
 import gov.nci.commonobjects.BreadCrumb;
-import gov.nci.commonobjects.Delighters;
 import com.relevantcodes.extentreports.LogStatus;
 
 public class BasicSearch_Test extends BaseClass {
